pyproject.toml
[tool.poetry]
name = "zurich_local"
version = "0.1.0"
description = ""
authors = ["lukas <stolz.lukas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.2.2"
duckdb = "^1.1.3"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
setuptools = "^75.6.0"
airbyte = "^0.21.0"
loguru = "^0.7.3"
pre-commit = "^4.0.1"
ruff = "^0.8.3"
isort = "^5.13.2"
dbt-duckdb = {extras = ["nmd"], version = "^1.9.1"}
st-pages = "^1.0.1"
folium = "^0.19.2"
streamlit-folium = "^0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
